+++ src/main/java/com/nifou/m/ifou_mobile_was/entity/result/Sub01Entity.java
@@ -34,7 +34,7 @@ public class Sub01Entity {
     private int totcnt;
     @Column
     @JsonProperty("amt")
-    private int totamt;
+    private Long totamt;
 
 
     public Sub01Entity() {
