+++ src/main/java/com/nifou/m/ifou_mobile_was/entity/result/Sub04Entity.java
@@ -39,7 +39,7 @@ public class Sub04Entity {
     private int totcnt;
     @Column
     @JsonProperty("amt")
-    private int totamt;
+    private Long totamt;
 
     public Sub04Entity() {
